[Refactor] refactor folder name
