--- src/main/java/com/capston/favicon/FaviconApplication.java
@@ -1,13 +1,0 @@
-package com.capston.favicon;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FaviconApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(FaviconApplication.class, args);
-	}
-
-}
