--- src/main/java/com/capston/favicon/user/application/service/MailService.java
@@ -1,5 +1,0 @@
-package com.capston.favicon.user.application.service;
-
-public interface MailService {
-    void send(String to, String subject, String text);
-}
