--- src/main/java/com/capston/favicon/user/application/service/OTPService.java
@@ -1,6 +1,0 @@
-package com.capston.favicon.user.application.service;
-
-public interface OTPService {
-    String generateOTP(String email);
-    boolean verifyOTP(String email, String otp);
-}
