--- src/main/java/com/capston/favicon/user/application/service/RedisService.java
@@ -1,8 +1,0 @@
-package com.capston.favicon.user.application.service;
-
-public interface RedisService {
-    void setCode(String email, String code);
-    String getCode(String email);
-
-    void deleteCode(String email);
-}
