--- src/main/java/com/capston/favicon/user/domain/User.java
@@ -1,22 +1,0 @@
-package com.capston.favicon.user.domain;
-
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.Getter;
-import lombok.Setter;
-
-
-@Getter
-@Setter
-@Entity
-@Table(name = "users")
-public class User {
-
-    @Id
-    private String email;
-    private String username;
-    private String password;
-
-}
