--- src/main/java/com/capston/favicon/user/dto/LoginDto.java
@@ -1,11 +1,0 @@
-package com.capston.favicon.user.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginDto {
-    private String email;
-    private String password;
-}
