--- src/main/java/com/capston/favicon/user/dto/SearchDto.java
@@ -1,8 +1,0 @@
-package com.capston.favicon.user.dto;
-
-import lombok.Getter;
-
-@Getter
-public class SearchDto {
-    private String text;
-}
