--- src/main/java/com/capston/favicon/user/repository/UserRepository.java
@@ -1,12 +1,0 @@
-package com.capston.favicon.user.repository;
-
-import com.capston.favicon.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-
-@Repository
-public interface UserRepository extends JpaRepository<User, String> {
-    User findByEmail(String email);
-}
-
