+++ src/main/java/com/capstone/favicon/FaviconApplication.java
@@ -1,0 +1,13 @@
+package com.capstone.favicon;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FaviconApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(FaviconApplication.class, args);
+	}
+
+}
