+++ src/main/java/com/capstone/favicon/config/SecurityConfig.java
@@ -1,0 +1,35 @@
+package com.capstone.favicon.config;
+
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+public class SecurityConfig {
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http
+                .csrf((csrfConfig) ->
+                        csrfConfig.disable()
+                )
+                .authorizeHttpRequests(authorize -> authorize
+                        .requestMatchers("/users/email-check", "/users/code-check", "users/register",
+                                "/users/login", "/users/logout",
+                                "/data-set/search-sorted", "/data-set/search-sorted/{category}",
+                                "/v3/api-docs/**", "/swagger-ui/**", "/swagger-ui.html").permitAll()
+                        .anyRequest().authenticated()
+                )
+                .httpBasic(httpBasic -> httpBasic.disable())
+                .formLogin(formLogin -> formLogin.disable())
+                .sessionManagement(session -> session
+                        .maximumSessions(1)
+                );
+
+        return http.build();
+    }
+}
