+++ src/main/java/com/capstone/favicon/user/application/DataServiceImpl.java
@@ -1,0 +1,30 @@
+package com.capstone.favicon.user.application;
+
+
+import com.capstone.favicon.user.application.service.DataService;
+import com.capstone.favicon.user.domain.Data;
+import com.capstone.favicon.user.repository.DataRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@RequiredArgsConstructor
+@Service
+public class DataServiceImpl implements DataService {
+
+    @Autowired
+    private DataRepository dataRepository;
+
+    @Override
+    public List<Data> search(String text) {
+        return dataRepository.searchByText(text);
+    }
+
+    @Override
+    public List<Data> searchWithCategory(String text, String category) {
+        return dataRepository.searchWithCategory(text, category);
+    }
+
+}
