+++ src/main/java/com/capstone/favicon/user/application/MailServiceImpl.java
@@ -1,0 +1,27 @@
+package com.capstone.favicon.user.application;
+
+import com.capstone.favicon.user.application.service.MailService;
+import org.springframework.mail.SimpleMailMessage;
+import org.springframework.mail.javamail.JavaMailSender;
+import org.springframework.stereotype.Service;
+
+
+@Service
+public class MailServiceImpl implements MailService {
+
+    private JavaMailSender mailSender;
+
+    public MailServiceImpl(JavaMailSender mailSender) {
+        this.mailSender = mailSender;
+    }
+
+    @Override
+    public void send(String to, String subject, String text) {
+        SimpleMailMessage message = new SimpleMailMessage();
+        message.setTo(to);
+        message.setSubject(subject);
+        message.setText(text);
+        mailSender.send(message);
+    }
+
+}
