+++ src/main/java/com/capstone/favicon/user/application/service/MailService.java
@@ -1,0 +1,5 @@
+package com.capstone.favicon.user.application.service;
+
+public interface MailService {
+    void send(String to, String subject, String text);
+}
