+++ src/main/java/com/capstone/favicon/user/application/service/OTPService.java
@@ -1,0 +1,6 @@
+package com.capstone.favicon.user.application.service;
+
+public interface OTPService {
+    String generateOTP(String email);
+    boolean verifyOTP(String email, String otp);
+}
