+++ src/main/java/com/capstone/favicon/user/application/service/RedisService.java
@@ -1,0 +1,8 @@
+package com.capstone.favicon.user.application.service;
+
+public interface RedisService {
+    void setCode(String email, String code);
+    String getCode(String email);
+
+    void deleteCode(String email);
+}
