+++ src/main/java/com/capstone/favicon/user/application/service/UserService.java
@@ -1,0 +1,13 @@
+package com.capstone.favicon.user.application.service;
+
+import com.capstone.favicon.user.dto.LoginDto;
+import com.capstone.favicon.user.dto.RegisterDto;
+import jakarta.servlet.http.HttpServletRequest;
+
+public interface UserService {
+    void sendCode(RegisterDto.checkEmail checkEmail);
+    void checkCode(RegisterDto.checkCode checkCode);
+    void join(RegisterDto registerDto);
+    void login(LoginDto loginDto, HttpServletRequest request);
+    void logout(HttpServletRequest request);
+}
