+++ src/main/java/com/capstone/favicon/user/dto/LoginDto.java
@@ -1,0 +1,11 @@
+package com.capstone.favicon.user.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginDto {
+    private String email;
+    private String password;
+}
