+++ src/main/java/com/capstone/favicon/user/dto/SearchDto.java
@@ -1,0 +1,8 @@
+package com.capstone.favicon.user.dto;
+
+import lombok.Getter;
+
+@Getter
+public class SearchDto {
+    private String text;
+}
