+++ src/main/java/com/capstone/favicon/user/repository/UserRepository.java
@@ -1,0 +1,12 @@
+package com.capstone.favicon.user.repository;
+
+import com.capstone.favicon.user.domain.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+
+@Repository
+public interface UserRepository extends JpaRepository<User, String> {
+    User findByEmail(String email);
+}
+
