--- src/test/java/com/capston/favicon/application/UserServiceTest.java
@@ -1,7 +1,0 @@
-package com.capston.favicon.application;
-
-
-public class UserServiceTest {
-
-
-}
