+++ src/test/java/com/capstone/favicon/application/UserServiceTest.java
@@ -1,0 +1,7 @@
+package com.capstone.favicon.application;
+
+
+public class UserServiceTest {
+
+
+}
